added new tests, reogranized tests, refactored gluonts predict
