+++ forecasting/src/autogluon/forecasting/models/__init__.py
@@ -1,1 +1,3 @@
 from .gluonts import AutoTabularModel, DeepARModel, SimpleFeedForwardModel, MQCNNModel
+
+__all__ = ["AutoTabularModel", "DeepARModel", "SimpleFeedForwardModel", "MQCNNModel"]
