--- forecasting/tests/unittests/models/test_abstract.py
@@ -1,4 +1,0 @@
-def test_module_importable():
-    from autogluon.forecasting.models.abstract import (  # noqa: F401
-        AbstractForecastingModel,
-    )
